% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasing_algorithm.R
\name{coupled_cpm_chains_Tblocking}
\alias{coupled_cpm_chains_Tblocking}
\title{Coupled correlated pseudo-marginal chains with blocking}
\usage{
coupled_cpm_chains_Tblocking(single_kernel, coupled_kernel, coupled_init,
  nparticles, nobs, iters_per_cycle = 1, K = 1, max_iterations = Inf,
  preallocate = 10, coupled_state = F, verbose = F, finish_time = NULL)
}
\arguments{
\item{single_kernel}{function taking a state (in a vector) and an iteration, and returning
a list with a key named \code{chain_state} and containing the next state.}

\item{coupled_kernel}{function taking two states (in two vectors) and an iteration,
and returning a list with keys \code{chain_state1} and \code{chain_state2}.}

\item{coupled_init}{function taking no arguments are returning an initial state for a Markov chain.}

\item{K}{number of iterations desired (will be proportional to the computing cost if meeting occurs before \code{K},
default to 1).}

\item{max_iterations}{number of iterations at which the function stops if it is still running  (default to Inf).}

\item{preallocate}{expected number of iterations, used to pre-allocate memory (default to 10).}
}
\description{
Sample two CPM chains, each following \code{single_kernel} marginally,
and \code{coupled_kernel} jointly, until min(max(tau, K), max_iterations), where tau
is the first time at which the two chains meet (i.e. take the same value exactly).
Or more precisely, they meet with a delay of one, i.e. X_t = Y_{t-1}. The chains
are initialized from the distribution provided in \code{rinit}.
}
