% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiasedpmcmc-package.R
\docType{package}
\name{debiasedpmcmc-package}
\alias{debiasedpmcmc-package}
\alias{debiasedpmcmc}
\title{debiasedpmcmc}
\description{
Unbiased particle MCMC estimators with couplings
}
\details{
This package contains scripts to remove the bias of particle MCMC estimators.
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>, Lawrence T. Middleton
}
\keyword{package}
