% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_kernel.R
\name{get_cpm_kernel}
\alias{get_cpm_kernel}
\title{Get correlated pseudo-marginal kernels}
\usage{
get_cpm_kernel(logtarget, Sigma_proposal, dimension, joint_update = T,
  rho = NULL, dim_U = NULL, component_updates = F, rho_component = NULL,
  block_indices_vec = NULL, single_kernel_only = F)
}
\arguments{
\item{logtarget}{function to compute target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{Sigma_proposal}{variance of the Normal random walk for the proposals}

\item{dimension}{dimension of the target distribution}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a function kernel
\code{kernel}
Optionally the kernels can perform blocked gibbs updates also
}
