% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_kernel.R
\name{get_cpm_T_blocking}
\alias{get_cpm_T_blocking}
\title{Get correlated pseudo-marginal kernels with T-blocking}
\usage{
get_cpm_T_blocking(logtarget, Sigma_proposal, dimension, rho_component,
  joint_update = T, rho = NULL, single_kernel_only = T)
}
\arguments{
\item{logtarget}{function to compute target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{Sigma_proposal}{variance of the Normal random walk for the proposals}

\item{dimension}{dimension of the target distribution}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a function kernel
\code{kernel}
Optionally the kernels can perform blocked gibbs updates also
}
