% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_kernel.R
\name{get_exchange_kernel}
\alias{get_exchange_kernel}
\title{Get exchange kernels}
\usage{
get_exchange_kernel(log_prior, d_log_lik, r_log_lik, Sigma_proposal, dimension)
}
\arguments{
\item{log_prior}{function to compute target log-prior}

\item{d_log_lik}{function to compute target likelihood log-density}

\item{r_log_lik}{function to sample target likelihood}

\item{Sigma_proposal}{variance of the Normal random walk for the proposals}

\item{dimension}{dimension of the target distribution}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the prior, likelihood densities and samplers
and tuning parameters, and returns a function kernel
\code{kernel}
}
