% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ising.R
\name{get_ising}
\alias{get_ising}
\title{Square Ising model}
\usage{
get_ising(M, J_lim, external_field = T, H_lim = NA, test_model = F)
}
\value{
A list
}
\description{
This function returns a list with objects as
* d_log_lik, r_lik, log_prior, three functions necessary to perform the exchange algorithm
}
