% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lgssm2params.R
\name{get_lgssm_2params}
\alias{get_lgssm_2params}
\title{Linear-gaussian state-space model (2 params)}
\usage{
get_lgssm_2params(dimension, sigma_y)
}
\value{
A list
}
\description{
This function returns a list with objects such as
* rinit, rinit_rand to sample from the initial distribution
* rtransition, rtransition_rand to sample from the transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* dimension, which represents the dimension of the latent process
}
