% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mv_probit.R
\name{get_mv_probit_model}
\alias{get_mv_probit_model}
\title{Multivariate probit model}
\usage{
get_mv_probit_model(init_mu, init_cov)
}
\value{
A list
}
\description{
This function returns a list with objects for random effects model
}
