% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_kernel.R
\name{get_pm_kernel}
\alias{get_pm_kernel}
\title{Get pseudo-marginal random walk Metropolis-Hastings kernels}
\usage{
get_pm_kernel(logtarget, state_crn_sample, latent_state, Sigma_proposal,
  dimension)
}
\arguments{
\item{logtarget}{function to compute target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{latent_state}{function to transform the random numbers into samples of the latent state}

\item{Sigma_proposal}{variance of the Normal random walk for the proposals}

\item{dimension}{dimension of the target distribution}

\item{state_crn_sampe}{function to sample the random numbers used to estimate the target}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a list containing the keys
\code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled pseudo marginal kernels, with Normal random walks.
These kernels can then be used in the function \code{\link{coupled_chains}}.
}
