% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_kernel.R
\name{get_pmmh_kernel_component_update}
\alias{get_pmmh_kernel_component_update}
\title{Get particle marginal Metropolis-Hastings kernel}
\usage{
get_pmmh_kernel_component_update(estimate_pftarget, Sigma_proposal, dimension)
}
\arguments{
\item{estimate_pftarget}{function to compute target log-density, e.g. see \code{\link{get_mvnormal}}}

\item{Sigma_proposal}{variance of the Normal random walk for the proposals}

\item{dimension}{dimension of the target distribution}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes descriptions of the target
and tuning parameters, and returns a function kernel
\code{kernel}
}
