% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pz.R
\name{get_pz_naturalparams}
\alias{get_pz_naturalparams}
\title{Phytoplankton-zooplankton model as in Jones, Parslow, Murray 2010}
\usage{
get_pz_naturalparams()
}
\value{
A list
}
\description{
The same as get_pz though implemented with natural params (without the transformation)
}
