% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_re.R
\name{get_re_model}
\alias{get_re_model}
\title{Random effects model of a stochastic frontier}
\usage{
get_re_model(init_mu, init_cov)
}
\value{
A list
}
\description{
This function returns a list with objects for random effects model
}
