% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghk.R
\name{ghk_obs_vectorise}
\alias{ghk_obs_vectorise}
\title{GHK algorithm (vectorised)}
\usage{
ghk_obs_vectorise(nsamp, y_obs, mean_mat, uniform_rvs, Sigma, K_units = NULL)
}
\arguments{
\item{nsamp}{number of particles x dimension of each observation}

\item{y_obs}{matrix of observations, nobs x dimension of each observation}

\item{mean_mat}{matrix of means for each observation in y_obs}

\item{uniform_rvs}{nobs x nsamp uniform random numbers}

\item{Sigma}{covariance matrix used in logistic regression}
}
\description{
runs the GHK algorithm to obtain importance sampling estimates for the observations
}
