% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{kf}
\alias{kf}
\title{kalman filter}
\usage{
kf(y, theta, mu_0, Sigma_0)
}
\arguments{
\item{y}{observation set}

\item{theta}{model parameters}

\item{mu_0}{initial mean}

\item{Sigma_0}{initial variance}
}
\description{
runs a kalman filter with specified model, model parameters
}
