% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{kf_1d}
\alias{kf_1d}
\title{kalman filter (efficient 1D implementation)}
\usage{
kf_1d(y, theta, mu_0, Sigma_0)
}
\arguments{
\item{y}{observation set}

\item{theta}{model parameters}

\item{mu_0}{initial mean}

\item{Sigma_0}{initial variance}
}
\description{
runs a kalman filter with specified model, model parameters
}
