% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{pf}
\alias{pf}
\title{bootstrap particle filter inference}
\usage{
pf(y, theta, model, nparticles, ess_threshold = 1, systematic = FALSE)
}
\arguments{
\item{y}{observation set}

\item{theta}{model parameters}

\item{model}{initial and transition densities for HMM}

\item{nparticles}{number of particles}
}
\description{
runs a particle filter with specified model, model parameters and number of particles
}
