% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{vec_gauss_max_coupling_mat}
\alias{vec_gauss_max_coupling_mat}
\title{Maximal coupling of two normals (vectorised code) where each row is maximally coupled with a diagonal covariance}
\usage{
vec_gauss_max_coupling_mat(mu1, mu2, s)
}
\arguments{
\item{mu1}{mean of first distribution}

\item{mu2}{mean of second distribution}

\item{s}{std dev of couplings}
}
\description{
Sample maximal couplings for gaussians with different means but same std. dev.
}
