% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_couplings.R
\name{vec_gauss_max_coupling}
\alias{vec_gauss_max_coupling}
\title{Maximal coupling of two normals (vectorised code)}
\usage{
vec_gauss_max_coupling(mu1, mu2, s)
}
\arguments{
\item{mu1}{mean of first distribution}

\item{mu2}{mean of second distribution}

\item{s}{std dev of couplings}
}
\description{
Sample maximal couplings for gaussians with different means but same std. dev.
}
